pyproject.toml
[project]
name = "gen-ai-testing-sytem-using-pytest"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.56.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
]
